[FIX] mrp: raise validation error if component is also byproduct
Having a product as component and byproduct causes an RPC error when opening the BoM overview. Odoo standard is expected to provide a validation error in this case to prevent the creation of such a BoM.

This change made test_10_variants inoperable so it was edited to use different components so as not to overlap with byproducts. This should be fine as the test in question only tests the creation of a BoM with variant specific components/work orders... and not the components themselves.